[Jim] Still working on Map will commit again after I make more progress

+++ src/backend/Coordinate.java
@@ -2,19 +2,19 @@ package backend;
 
 public class Coordinate {
 	
-	private int x;
-	private int y;
+	private double x;
+	private double y;
 	
-	public Coordinate(int x, int y) {
+	public Coordinate(double x, double y) {
 		this.x = x;
 		this.y = y;
 	}
 
-	public int getX() {
+	public double getX() {
 		return this.x;
 	}
 	
-	public int getY() {
+	public double getY() {
 		return this.y;
 	}
 }
